requirements.txt
bcrypt==3.2.2
cffi==1.17.1
Faker==37.3.0
mysql-connector-python==9.3.0
passlib==1.7.4
pycparser==2.22
python-dotenv==1.1.0
tzdata==2025.2
